requirements.txt
#torch==2.9.1+cpu
#torchvision==0.24.1+cpu
torch==2.9.1
torchvision==0.24.1
typer==0.21.1
fastapi==0.128.0
uvicorn==0.40.0
matplotlib>=3.10.0
dvc[gdrive,gs]==3.66.1
hydra-core==1.3.2
torch-tb-profiler==0.4.3
loguru==0.7.3
wandb==0.24.0
scikit_learn==1.8.0
Pillow==12.1.0
google-cloud-storage==3.8.0
python-dotenv==1.0.1
coverage==7.13.1
snakeviz==2.2.0
tensorboard==2.20.0
prometheus-client==0.24.1
pytorch-lightning==2.6.0

# FE
requests==2.32.5
httpx==0.28.1
locust==2.43.1
bentoml==1.4.33
